Handle @. and %. attributes, too!
Bind values into the hash instead of assigning them so that simple
array and hash attributes ("has Int @.stuff" instead of
"has Array[Int] $.stuff") may be initialized correctly.

Add a bit of heuristics to initialize such attributes using the Array
and Hash classes instead of their "actual" types, since their "actual"
types are reported as the Positional and Associative roles which may
not really be constructed in this way.

Strip off @. and %. in the attribute names, too.

Remove the BUGS section in the documentation, we just fixed it! :)

Add the 02-twigils test, exactly the same as the 01-simple one, but
using twigils instead of Array[].

Bump the distribution version to 0.2.0.

Suggested by:		<jnthn@jnthn.net> in RT #127315